Generalize Crypto module
