+++ pkg/modules/modules.go
@@ -21,7 +21,6 @@ type Module interface {
 // The Modules structs groups the modules a Node consists of.
 type Modules struct {
 	Net           Net              // Sends messages produced by Mir through the network.
-	Crypto        Crypto           // Performs cryptographic operations (except for computing hashes)
 	WAL           WAL              // Implements a persistent write-ahead log for the case of crashes and restarts.
 	ClientTracker ClientTracker    // Keeps the state related to clients and validates submitted requests.
 	RequestStore  RequestStore     // Provides persistent storage for request data.
